+++ req.in
@@ -12,6 +12,7 @@ pandas==1.4.4
 plotly==5.10.0
 rich
 keyboard
+perlin_noise
 
 
 # Для типизации
