+++ req.txt
@@ -129,13 +129,8 @@ pandocfilters==1.5.0
     # via nbconvert
 parso==0.8.3
     # via jedi
-<<<<<<< HEAD:requirements.txt
-pathspec==0.10.1
-    # via black
 perlin-noise==1.12
-    # via -r requirements.in
-=======
->>>>>>> main:req.txt
+    # via -r req.in
 pickleshare==0.7.5
     # via ipython
 plotly==5.10.0
