+++ src/labs/lab2/__init__.py
@@ -1,0 +1,6 @@
+from src.labs.lab2.field import Field
+from src.labs.lab2.unit import Unit
+from src.labs.lab2.environment import Environment
+
+from typing import Tuple
+Pos = Tuple[int, int]
